git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@40649 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ libpcsxcore/ppf.c
@@ -91,7 +91,7 @@ void FreePPFCache() {
 	ppfCache = NULL;
 }
 
-void CheckPPFCache(unsigned char *pB, int m, int s, int f) {
+void CheckPPFCache(unsigned char *pB, u8 m, u8 s, u8 f) {
 	PPF_CACHE *pcstart, *pcend, *pcpos;
 	int addr = MSF2SECT(btoi(m), btoi(s), btoi(f)), pos, anz, start;
